seat: move uterm monitor to seat
